change bullet sim

+++ contracts/src/types.cairo
@@ -19,8 +19,8 @@ struct TurnMove {
 
 #[derive(Copy, Drop, Serde, Introspect)]
 struct IVec2 {
-    x: u32,
-    y: u32,
+    x: u64,
+    y: u64,
     xdir: bool,
     ydir: bool
 }
